+++ tools/eubfr-cli/commands/services/delete.js
