+++ tools/eubfr-cli/commands/services/deploy.js
